+++ src/core/kernel/radiance.hpp
@@ -13,8 +13,8 @@ namespace core
 {
 template <typename Random>
 PolyFunction( CosSampleHemisphere, Require<Random> )(
-  ()->double3 {
-	  double3 w;
+  ()->float3 {
+	  float3 w;
 	  auto r1 = 2 * M_PI * call<Random>(), r2 = call<Random>(), r2s = sqrt( r2 );
 
 	  return w;
@@ -22,20 +22,20 @@ PolyFunction( CosSampleHemisphere, Require<Random> )(
 
 // template <typename Random>
 // PolyFunction( SampleBsdf, Require<Random> )(
-//   ( const double3 &wo )->double3 {
-// 	  double3 w;
+//   ( const float3 &wo )->float3 {
+// 	  float3 w;
 // 	  auto r1 = 2 * M_PI * call<Random>(), r2 = call<Random>(), r2s = sqrt( r2 );
-// 	  auto u = normalize( cross( abs( wo.x ) > abs( wo.y ) ? double3{ 0, 1, 0 } : double3{ 1, 0, 0 }, wo ) );
+// 	  auto u = normalize( cross( abs( wo.x ) > abs( wo.y ) ? float3{ 0, 1, 0 } : float3{ 1, 0, 0 }, wo ) );
 // 	  auto v = cross( wo, u );
 // 	  return normalize( ( u * cos( r1 ) + v * sin( r1 ) ) * r2s + wo * sqrt( 1 - r2s ) );
 //   } );
 
 template <typename Random>
 PolyFunction( Radiance, Require<Random> )(
-  ( const Ray &r, const Scene &scene, Allocator &pool )->double3 {
+  ( const Ray &r, const Scene &scene, Allocator &pool )->float3 {
 	  auto ray = r;
 	  Interreact isect;
-	  double3 L = { 0, 0, 0 };
+	  float3 L = { 0, 0, 0 };
 	  constexpr auto maxBounce = 10;
 
 	  for ( auto bounce = 0; ( isect = scene.intersect( ray, pool ) ) && bounce != maxBounce; ++bounce )
