+++ src/core/meta/bsdf.hpp
@@ -13,12 +13,12 @@ struct BxDF
 {
 	KOISHI_HOST_DEVICE virtual ~BxDF() = default;
 
-	KOISHI_HOST_DEVICE virtual double3 f( const double3 &wo, const double3 &wi ) const = 0;
-	KOISHI_HOST_DEVICE virtual double pdf( const double3 &wo, const double3 &wi ) const  // cosHemisphere
+	KOISHI_HOST_DEVICE virtual float3 f( const float3 &wo, const float3 &wi ) const = 0;
+	KOISHI_HOST_DEVICE virtual float pdf( const float3 &wo, const float3 &wi ) const  // cosHemisphere
 	{
 		return hemisphere::isSame( wo, wi ) ? hemisphere::h( wi ) * invPI : 0.;
 	}
-	KOISHI_HOST_DEVICE virtual double3 sample_f( const double3 &wo, double3 &wi, const double2 &rn, double &pdf ) const
+	KOISHI_HOST_DEVICE virtual float3 sample_f( const float3 &wo, float3 &wi, const float2 &rn, float &pdf ) const
 	{
 		wi = hemisphere::sampleCos( rn );
 		return f( wo, wi );
@@ -34,12 +34,12 @@ struct BSDF final
 		bxdfs[ numBxdfs++ ] = create<T>( pool );
 	}
 
-	KOISHI_HOST_DEVICE double3 f( const double3 &wo, const double3 &wi ) const
+	KOISHI_HOST_DEVICE float3 f( const float3 &wo, const float3 &wi ) const
 	{
 		// Vector3f wi = WorldToLocal( wiW ), wo = WorldToLocal( woW );
 		// if ( wo.z == 0 ) return 0.;
 		// bool reflect = Dot( wiW, ng ) * Dot( woW, ng ) > 0;
-		double3 f = { 0, 0, 0 };
+		float3 f = { 0, 0, 0 };
 		// for ( int i = 0; i < nBxdfs; ++i )
 		// 	if ( bxdfs[ i ]->MatchesFlags( flags ) &&
 		// 		 ( ( reflect && ( bxdfs[ i ]->type & BSDF_REFLECTION ) ) ||
@@ -47,10 +47,10 @@ struct BSDF final
 		// 		f += bxdfs[ i ]->f( wo, wi );
 		return f;
 	}
-	KOISHI_HOST_DEVICE double pdf( const double3 &wo, const double3 &wi ) const
+	KOISHI_HOST_DEVICE float pdf( const float3 &wo, const float3 &wi ) const
 	{
 	}
-	KOISHI_HOST_DEVICE double3 sample_f( const double3 &wo, double3 &wi, const double2 &rn, double &pdf ) const
+	KOISHI_HOST_DEVICE float3 sample_f( const float3 &wo, float3 &wi, const float2 &rn, float &pdf ) const
 	{
 		int comp = min( (int)floor( rn.x * numBxdfs ), (int)numBxdfs - 1 );
 		auto bxdf = bxdfs[ comp ];
